requirements.txt
aniso8601==9.0.1
azure-core==1.30.2
azure-storage-blob==12.20.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-RESTful==0.3.10
hurry.filesize==0.9
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
pycparser==2.22
pytz==2024.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3